Add initial dependency injection
