+++ src/electionguard_gui/services/db_service.py
@@ -4,12 +4,21 @@ from electionguard_gui.services.configuration_service import (
     get_db_host,
     get_db_password,
 )
+from electionguard_gui.services.eel_log_service import EelLogService
 
 from electionguard_gui.services.service_base import ServiceBase
+from dependency_injector.wiring import Provide, inject
 
 
 class DbService(ServiceBase):
     """Responsible for instantiating a database"""
+
+    log_service: EelLogService
+
+    def __init__(self, log_service: EelLogService) -> None:
+        self.log_service = log_service
+        self._db_password = get_db_password()
+        self._db_host = get_db_host(self.DEFAULT_HOST)
 
     DEFAULT_HOST = "localhost"
     DEFAULT_PORT = 27017
@@ -17,10 +26,6 @@ class DbService(ServiceBase):
 
     _db_password: str
     _db_host: str
-
-    def __init__(self) -> None:
-        self._db_password = get_db_password()
-        self._db_host = get_db_host(self.DEFAULT_HOST)
 
     def get_db(self) -> Database:
         client: MongoClient = MongoClient(
